+++ lib/sg_strange_calendar/presenter/horizonal_presenter.rb
@@ -1,33 +1,9 @@
 require 'date'
+require_relative 'base_presenter'
 
 class SgStrangeCalendar
   class Presenter
-    class HorizonalPresenter
-      def self.present(array, today)
-        array.map do |row|
-          date = row[1..].map do |cell|
-            display_value(cell, today)
-          end.join(' ')
-          # row[0] is left-justified
-          # row[1..] is right-justified
-          # rsrip is to remove trailing spaces
-          str_row = [
-            display_value(row[0], today).ljust(4),
-            date
-          ].join(' ').rstrip
-          # [xx] を含む row は今日を含んでいると判断する
-          if str_row.match?(/[\[|\]]/)
-            # 右側へのはみ出しを削る
-            str_row = str_row.match?(/\] /) ? str_row.gsub(/\] /, ']') : str_row
-            # 2桁以上の場合は、左側へのはみ出しを削る
-            if today.day > 9
-              str_row = str_row.match?(/ \[/) ? str_row.gsub(/ \[/, '[') : str_row
-            end
-          end
-          str_row
-        end.join("\n")
-      end
-
+    class HorizonalPresenter < BasePresenter
       class << self
         def display_value(value, today)
           # for debug
@@ -39,9 +15,15 @@ class SgStrangeCalendar
           value.to_s.rjust(2)
         end
 
-        def is_today?(value, today)
-          false if value.is_a?(Date)
-          value == today
+        def update_today_str(str_row, today)
+          # 右側へのはみ出しを削る
+          str_row = str_row.match?(/\] /) ? str_row.gsub(/\] /, ']') : str_row
+          # 2桁以上の場合は、左側へのはみ出しを削る
+          if today.day > 9
+            str_row = str_row.match?(/ \[/) ? str_row.gsub(/ \[/, '[') : str_row
+          end
+
+          str_row
         end
       end
     end
